audit(training): Wave 3 MEDIUM — 6 parallel fix-agents, +1740 LOC source, 36 new regression tests
6 fix-agents (composite, core, feature_handling, neural, eval, helpers)
applied MEDIUM-tier audit findings to mlframe/training/. Tracking
again dropped before final reports, but progress checkpoints under
audit/progress/ confirm 30+ fixes landed; verified post-hoc via
pytest on Wave 3 + previous waves: 91/91 audit tests green.

Wave 3 source highlights (~30 MEDIUM fixes across 47 files):
- composite: FIFO->true LRU in OOF_HOLDOUT + refit_cache via
  OrderedDict.move_to_end; argsort-of-argsort -> scipy.stats.rankdata
  for tie-handled Spearman in feature-type demoter; duplicate
  deque import removed; stdlib imports promoted to module level in
  composite_cache; NotFittedError unified across feature_importances_
  / coef_ / get_booster / booster_ / n_features_in_ getters
- core: math.isnan replaces NaN-as-mismatch check in
  _phase_finalize; explicit-dict dispatch replaces locals() trick
  in _phase_helpers polars-frame merge loop
- feature_handling: providers._is_secret_field uses whole-token
  match (api_key matches, tokenizer does NOT); _scrub_dict recurses
  into nested dicts + list-of-dicts (Authorization header value
  was leaking through model_dump pre-fix); from_uri URL-decodes
  model name (BAAI%2Fbge -> BAAI/bge); cache_backend._evict_to_caps
  fast-path skips listdir+stat when entry-count well under cap;
  fingerprint cache key uses weakref-style id+strong-ref to avoid
  GC-id reuse; custom_handler narrows TypeError catch to AttributeError;
  text_detection.pl.Enum dtype check via isinstance against the
  current polars datatype class; registry.prewarm clears future on
  exception
- neural: pack_padded_sequence zero-length clamp; R2Score
  single-sample NaN guard; DataLoader generator seeding for
  reproducibility; ThreadPool threshold lifted to >100k seqs;
  xxhash import promoted to module level; sequence_preprocessing
  helper imports hoisted out of hot loop; flat.py argmax fix for
  multilabel preds; cache key uses content-hash not 3-float
  sampling; ranker set_epoch wired via Lightning callback; XGBRanker
  EarlyStopping moved to fit_kwargs; CatBoost init_kwargs ordering
  preserves user verbose; ranker_suite ensemble fallthrough WARNs
  before silent rrf fallback
- eval: dummy_baselines / _dummy_baseline_compute / _eval_helpers
  / drift_report / target_temporal_audit / regression_residual_audit
  / evaluation: sweep S1 (dated audit-history comments stripped),
  S2 (em-dash -> ' - ' in log strings), S3 (assert -> if raise on
  module-boundary validation), period_freq updated for pandas 2.x
  (M -> ME, Q -> QE, Y -> YE)
- helpers: utils.coerce_to_numpy fast-path for pandas/polars Series;
  _cb_pool._GPU_PROBE_LOCK + _ram_helpers._MAYBE_CLEAN_LOCK make
  module-level cached state thread-safe; io.load_mlframe_model
  uses logger.exception so traceback survives; lgb_shim eval_set
  list-of-matrices guard; helpers.get_trainset_features_stats
  pandas vectorise (no per-col Python iter); mlp_runtime_defaults
  device_id param for multi-GPU; splitting factorize-based date
  label LUT replaces per-row dict.map(); helpers.py noqa F401
  re-export of underscore-private names dropped; _classif_helpers
  _ChainEnsemble param defaults reverted to constructor signature

Wave 3 regression tests added:
- tests/training/test_audit_2026_05_17_medium_composite.py (9)
- tests/training/test_audit_2026_05_17_medium_feature_handling.py (14)
- tests/training/test_audit_2026_05_17_medium_neural.py (13)

Helpers / core / eval test files: agents lost tracking before
writing them; the source fixes still land via this commit. Test
backfill deferred to follow-up.

Post-commit verification: 91/91 audit-tagged tests pass:
- Wave 1 critical: 12/12
- Wave 1.5 factory refactor: covered in critical
- Wave 2 high (eval/feature_handling/helpers/neural): 43/43
- Wave 3 medium (composite/feature_handling/neural): 36/36

audit/progress/ checkpoints retained for traceability of
agent-side work. Full atomic-finding disposition table at
audit/CODE_REVIEW_2026-05-17.md.

+++ src/mlframe/training/pu_learning.py
@@ -355,13 +355,18 @@ class PULearningWrapper(BaseEstimator, ClassifierMixin):
                     "unbiased negative). Found none -- switch to "
                     "unbiased_only or check is_unbiased flag."
                 )
-            sample_weight = np.where(s == 1, 1.0 / n1, 1.0 / n0).astype(np.float32)
+            # float64 throughout: 1.0/n1 for large n (>= 2**24) loses precision
+            # in float32 and the per-class balance multiplier compounds the
+            # rounding error. Most sklearn estimators internally coerce
+            # sample_weight to float64 anyway, so producing float32 here only
+            # paid the rounding cost without saving downstream work.
+            sample_weight = np.where(s == 1, 1.0 / n1, 1.0 / n0).astype(np.float64)
             sample_weight *= (n0 + n1)  # rescale for numerical stability
         else:
             sample_weight = None
 
         if "sample_weight" in fit_params:
-            extra_sw = np.asarray(fit_params.pop("sample_weight"), dtype=np.float32)
+            extra_sw = np.asarray(fit_params.pop("sample_weight"), dtype=np.float64)
             sample_weight = (sample_weight * extra_sw) if sample_weight is not None else extra_sw
 
         self.base_estimator_ = clone(self.base_estimator)
